pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "AAPL 10-K research agent with Milvus hybrid retrieval"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LangChain 1.0 LTS — strict semver, allow minor updates
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langchain-openai>=0.3.0",
    "langchain-milvus>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    # LangGraph 1.0 LTS
    "langgraph>=1.0,<2.0",
    # Observability
    "langsmith>=0.3.0",
    # Runtime
    "python-dotenv>=1.0.1",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/api"]

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.1.0",
    "pytest>=8.3.5",
    "ruff>=0.8.0",
]
